Merge branch 'cordova-sqlite-legacy-core' into storage-master

+++ spec/www/spec/db-open-close-delete-test.js
@@ -1485,16 +1485,46 @@ var mytests = function() {
       });
     }
 
-    describe('repeated open/close/delete test(s)', function() {
-      var scenarioName = isAndroid ? 'Plugin-implementation-default' : 'Plugin';
-      var suiteName = scenarioName + ': ';
+    for (var i=0; i<pluginScenarioCount; ++i) {
+
+      describe(pluginScenarioList[i] + ': repeated open/close/delete test(s)', function() {
+        var scenarioName = pluginScenarioList[i];
+        var suiteName = scenarioName + ': ';
+        var isImpl2 = (i === 1);
 
         // NOTE: MUST be defined in function scope, NOT outer scope:
         var openDatabase = function(first, second, third) {
+          if (first.constructor === String ) throw new Error('string not expected here');
+
+          // androidDatabaseImplementation: 2 (builtin android.database implementation):
+          if (isImpl2) {
+            var dbname = first.name;
+            return window.sqlitePlugin.openDatabase({
+              name: 'i2-'+dbname,
+              // database location setting needed here (value ignored on Android):
+              location: 'default',
+              androidDatabaseImplementation: 2,
+              androidLockWorkaround: 1
+            }, second, third);
+          }
+
           return window.sqlitePlugin.openDatabase(first, second, third);
         }
 
         var deleteDatabase = function(first, second, third) {
+          if (first.constructor === String ) throw new Error('string not expected here');
+
+          // androidDatabaseImplementation: 2 (builtin android.database implementation):
+          if (isImpl2) {
+            var dbname = first.name;
+            return window.sqlitePlugin.deleteDatabase({
+              name: 'i2-'+dbname,
+              // database location setting needed here (value ignored on Android):
+              location: 'default',
+              androidDatabaseImplementation: 2
+            }, second, third);
+          }
+
           window.sqlitePlugin.deleteDatabase(first, second, third);
         }
 
@@ -1509,23 +1539,23 @@ var mytests = function() {
           var db1 = openDatabase(dbargs, function () {
             var db2 = openDatabase(dbargs, function () {
               db1.readTransaction(function(tx1) {
-                tx1.executeSql('SELECT 1', [], function(tx1d, results) {
+                tx1.executeSql('SELECT 1', [], function(tx_ignored, results) {
                   ok(true, 'db1 transaction working');
                   start(1);
-                }, function(ignored, error) {
+                }, function(tx_ignored, error) {
                   ok(false, error);
                 });
               }, function(error) {
                 ok(false, error);
               });
               db2.readTransaction(function(tx2) {
-                tx2.executeSql('SELECT 1', [], function(tx2d, results) {
+                tx2.executeSql('SELECT 1', [], function(tx_ignored, results) {
                   ok(true, 'db2 transaction working');
                   start(1);
-                }, function(ignored, error) {
+                }, function(tx_ignored, error) {
                   ok(false, error);
                 });
-              }, function(error) {
+              }, function(tx_ignored, error) {
                 ok(false, error);
               });
             }, function (error) {
@@ -1779,6 +1809,8 @@ var mytests = function() {
         test_it(suiteName + ' repeatedly open and close database faster (5x)', function () {
           // TBD CURRENTLY BROKEN on iOS/macOS due to current background processing implementation:
           if (!isAndroid && !isWindows && !isWP8) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          // TBD ???:
+          if (isAndroid && isImpl2) pending('FAILS on builtin android.database implementation (androidDatabaseImplementation: 2)');
 
           var dbName = 'repeatedly-open-and-close-faster-5x.db';
           var dbargs = {name: dbName, location: 'default'};
@@ -1958,7 +1990,8 @@ var mytests = function() {
           });
         });
 
-    });
+      });
+    }
 
   });
